# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation API",
        "description": "Documentation for API.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/matter/{matterId}": {
            "get": {
                "tags": [
                    "Matter"
                ],
                "summary": "Get details by matter ID",
                "operationId": "1d0f941cdc8e1faae314b4cbff1016d5",
                "parameters": [
                    {
                        "name": "matterId",
                        "in": "path",
                        "description": "Matter ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "MatterNumber": {
                                                "type": "string",
                                                "example": "331491"
                                            },
                                            "MatterDesc": {
                                                "type": "string",
                                                "example": "Purchase of 22 PR AZURE TEST, Bangor"
                                            },
                                            "MatterOpenedDate": {
                                                "type": "string",
                                                "example": "2017-09-21"
                                            },
                                            "CaseHandler": {
                                                "type": "string",
                                                "example": "ASUMMERS"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/matter/{matterId}/single": {
            "get": {
                "tags": [
                    "Matter"
                ],
                "summary": "Get single matter by Id",
                "operationId": "8841ab11ff7397beffc1644787219a99",
                "parameters": [
                    {
                        "name": "matterId",
                        "in": "path",
                        "description": "Matter ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "MatterNumber": {
                                                "type": "string",
                                                "example": "331491"
                                            },
                                            "MatterDesc": {
                                                "type": "string",
                                                "example": "Purchase of 22 PR AZURE TEST, Bangor"
                                            },
                                            "MatterOpenedDate": {
                                                "type": "string",
                                                "example": "2017-09-21"
                                            },
                                            "CaseHandler": {
                                                "type": "string",
                                                "example": "ASUMMERS"
                                            },
                                            "CaseHandlerName": {
                                                "type": "string",
                                                "example": "Adele Summers"
                                            },
                                            "CaseHandlerTel": {
                                                "type": "string",
                                                "example": "028 9032 1234"
                                            },
                                            "CaseHandlerPhotograph": {
                                                "type": "string",
                                                "example": "aa...bb"
                                            },
                                            "CaseHandlerEmail": {
                                                "type": "string",
                                                "example": "ASummers@wilson-nesbitt.co.uk"
                                            },
                                            "Client": {
                                                "type": "string",
                                                "example": "Mr Alex Bittles and Mrs Jane Bloggs"
                                            },
                                            "Progress": {
                                                "type": "string",
                                                "example": "80.0"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/matter/{matterId}/history": {
            "get": {
                "tags": [
                    "Matter"
                ],
                "summary": "Get matter history by Id",
                "operationId": "b1897cf3401d5a5cc98f2262cfc081f6",
                "parameters": [
                    {
                        "name": "matterId",
                        "in": "path",
                        "description": "Matter ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "HistoryDate": {
                                                "type": "string",
                                                "example": "2024-10-24 15:18:48.000"
                                            },
                                            "HistoryDesc": {
                                                "type": "string",
                                                "example": "File in BOX Reference: BAN-2024-CON-111"
                                            },
                                            "HistoryNotes": {
                                                "type": "string",
                                                "example": ""
                                            },
                                            "HistoryCategory": {
                                                "type": "string",
                                                "example": ""
                                            },
                                            "HistoryID": {
                                                "type": "string",
                                                "example": "21779964"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/matter/{matterId}/milestone-dates": {
            "get": {
                "tags": [
                    "Matter"
                ],
                "summary": "Get matter milestones dates by Id",
                "operationId": "5aea4a5c047d52b9a53396e23ef7e6f7",
                "parameters": [
                    {
                        "name": "matterId",
                        "in": "path",
                        "description": "Matter ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "Milestone": {
                                                "type": "string",
                                                "example": "Instructions Confirmed by Client"
                                            },
                                            "CompletedDate": {
                                                "type": "string",
                                                "example": "2022-08-08"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}